Charuza_Kuśmierz_Sroczyk/hallucinations/pyproject.toml
[project]
name = "nlp-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.124.4",
    "numpy==2.4.0rc1",
    "pandas==3.0.0rc0",
    "python-dotenv==1.2.1",
    "torch>=2.9.1",
    "transformers==5.0.0rc1",
    "uvicorn>=0.38.0",
    "pydantic>=2.7.0",
    "huggingface_hub>=0.25.2",
    "google-genai>=1.55.0",
    "tqdm",
    "openai>=1.3.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]
